+++ NEWS.md
@@ -1,3 +1,5 @@
+# vroom (development version)
+
 # vroom 1.5.4
 
 * `vroom(col_types=)` now accepts column type names like those accepted by utils::read.table. e.g.